backend/requirements.txt
# ============================================================
# Lead Qualifier — Python Dependencies
# Install: pip install -r requirements.txt
# Then:    playwright install chromium
# ============================================================

# --- Web Scraping ---
crawl4ai>=0.3.0
playwright>=1.40.0

# --- AI / LLM ---
openai>=1.0.0          # OpenAI + Kimi (OpenAI-compatible API)
tiktoken>=0.5.0        # Token counting for cost estimation

# --- Lead Discovery (optional, for test_exa.py) ---
exa-py>=1.0.0          # Exa AI neural search

# --- Data Validation ---
pydantic>=2.0.0
pandas>=2.0.0

# --- Environment ---
python-dotenv>=1.0.0

# --- CLI / Progress ---
tqdm>=4.65.0
rich>=13.0.0

# --- Image Processing ---
Pillow>=10.0.0         # Screenshot resizing for vision API

# --- HTTP ---
httpx>=0.25.0          # Async HTTP (enrichment APIs)
aiohttp>=3.9.0
asyncio-throttle>=1.0.0

# --- Chat API Server ---
fastapi>=0.115.0       # API framework for chat interface
uvicorn[standard]>=0.30.0  # ASGI server

# --- Auth & Database ---
pyjwt[crypto]>=2.8.0    # JWT verification with JWKS support (RS256/ES256)
sqlalchemy>=2.0.0      # ORM
aiosqlite>=0.19.0      # Async SQLite driver (local dev fallback)
asyncpg>=0.29.0        # Async PostgreSQL driver (Supabase)
alembic>=1.12.0        # Database migrations

# --- Testing ---
pytest>=8.0.0
pytest-asyncio>=0.24.0
